Replace `retrieve` with `retrieveHourly`
