+++ api/src/routes/index.js
@@ -1,5 +1,4 @@
 import retrieve from './retrieve.js';
-import retrieveHourly from './retrieve-hourly.js';
 
 export default (app) => {
   app.get('/', (req, res) => {
@@ -7,5 +6,4 @@ export default (app) => {
   });
 
   app.get('/retrieve', retrieve());
-  app.get('/retrieve-hourly', retrieveHourly());
 };
